+++ src/essential-select/essential-select-config.ts
@@ -1,0 +1,13 @@
+export class EssentialSelectModuleConfig {
+
+    /**
+     * Default language if window.navigator is empty
+     */
+    defaultLanguage?: string;
+
+    /**
+     * Force language to use instead if window.navigator
+     */
+    forcedDefaultLanguage?: string;
+
+}